new-requirements.txt
# annotated-types==0.7.0
# anyio==4.9.0
# attrs==25.3.0
# backoff==2.2.1
# bcrypt==4.3.0
# blinker==1.9.0
# build==1.2.2.post1
# cachetools==5.5.2
# certifi==2025.7.14
# charset-normalizer==3.4.2
# chromadb==1.0.15
# click==8.2.1
# colorama==0.4.6
# coloredlogs==15.0.1
# distro==1.9.0
# durationpy==0.10
# filelock==3.18.0
# Flask==3.1.1
# flask-cors==6.0.1
# flatbuffers==25.2.10
# fsspec==2025.7.0
# google-auth==2.40.3
# googleapis-common-protos==1.70.0
# grpcio==1.74.0
# h11==0.16.0
# httpcore==1.0.9
# httptools==0.6.4
# httpx==0.28.1
# huggingface-hub==0.34.1
# humanfriendly==10.0
# idna==3.10
# importlib_metadata==8.7.0
# importlib_resources==6.5.2
# itsdangerous==2.2.0
# Jinja2==3.1.6
# jiter==0.10.0
# jsonschema==4.25.0
# jsonschema-specifications==2025.4.1
# kubernetes==33.1.0
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# mdurl==0.1.2
# mmh3==5.1.0
# mpmath==1.3.0
# numpy==2.3.2
# oauthlib==3.3.1
# onnxruntime==1.22.1
# openai==1.97.1
# opentelemetry-api==1.35.0
# opentelemetry-exporter-otlp-proto-common==1.35.0
# opentelemetry-exporter-otlp-proto-grpc==1.35.0
# opentelemetry-proto==1.35.0
# opentelemetry-sdk==1.35.0
# opentelemetry-semantic-conventions==0.56b0
# orjson==3.11.1
# overrides==7.7.0
# packaging==25.0
# pandas==2.3.1
# posthog==5.4.0
# protobuf==6.31.1
# pyasn1==0.6.1
# pyasn1_modules==0.4.2
# pybase64==1.4.1
# pydantic==2.11.7
# pydantic_core==2.33.2
# Pygments==2.19.2
# PyPika==0.48.9
# pyproject_hooks==1.2.0
# pyreadline3==3.5.4
# python-dateutil==2.9.0.post0
# python-dotenv==1.1.1
# pytz==2025.2
# PyYAML==6.0.2
# referencing==0.36.2
# requests==2.32.4
# requests-oauthlib==2.0.0
# rich==14.1.0
# rpds-py==0.26.0
# rsa==4.9.1
# shellingham==1.5.4
# six==1.17.0
# sniffio==1.3.1
# sympy==1.14.0
# tenacity==9.1.2
# tokenizers==0.21.2
# tqdm==4.67.1
# typer==0.16.0
# typing-inspection==0.4.1
# typing_extensions==4.14.1
# tzdata==2025.2
# urllib3==2.5.0
# uvicorn==0.35.0
# watchfiles==1.1.0
# websocket-client==1.8.0
# websockets==15.0.1
# Werkzeug==3.1.3
# zipp==3.23.0



# Flask==2.3.3
# Flask-CORS==4.0.0
# chromadb==0.4.15
# openai==0.28.1
# requests==2.31.0
# numpy==1.24.3
# pandas==2.0.3
# python-dotenv==1.0.0

# pypdf2==3.0.1
# python-docx==0.8.11
# openpyxl==3.1.2
# python-magic-bin==0.4.14
# langchain==0.1.0
# langchain-text-splitters==0.0.1



# Core Flask and API
Flask==3.1.1
Flask-CORS==6.0.1
Werkzeug==3.1.3

# ChromaDB and Vector Store
chromadb==1.0.15

# OpenAI
openai==1.97.1

# Data Processing
numpy==2.3.2
pandas==2.3.1

# Environment and Config
python-dotenv==1.1.1

# File Processing
pypdf2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5

# Text Processing - Simplified versions
langchain-core==0.3.34
langchain-text-splitters==0.3.4

# Requests
requests==2.32.4

# Additional dependencies
pydantic==2.11.7
typing-extensions==4.14.1